Clean mainInfo collecting information being done inside mainInfo
